--- src/models/__init__.py
@@ -1,11 +1,0 @@
-from models.birthday import Birthday
-from models.field import Field
-from models.phone import Phone
-from models.name import Name
-from models.address import Address
-from models.note import Note
-from models.record import Record
-from models.address_book import AddressBook
-
-
-__all__ = ["Birthday", "Field", "Phone", "Record", "Name", "Address","Note", "AddressBook", "Email"]
