+++ src/models/name.py
@@ -1,4 +1,4 @@
-from models import Field
+from models.field import Field
 
 
 class Name(Field):
